Move classes Service and Server to module
